Move code from stats controller to stats handler
